hello a lot of stuff was updated like the jquery transitions and buttons
