+++ frontend/assets/stylesheets/base.scss
@@ -33,6 +33,20 @@ body {
   background-color: #F9F7F9;
 }
 
+.htmlbg {
+  background-color: #F9F7F9;
+}
+
+.edit-background-in {
+  background-color:#6190E8 !important;
+  transition:all ease-in-out 0.6s;
+}
+
+.edit-background-out {
+  background-color:#F9F7F9 !important;
+  transition:all ease-in-out 0.6s;
+}
+
 .smallfade-appear {
   opacity: 0.5;
 }
