+++ frontend/components/clubProfile/CompetitionRecord.js
@@ -29,8 +29,8 @@ class CompetitionRecord extends Component {
   }
   render() {
     return(
-      <div className='tile is-parent is-vertical' style={{backgroundColor: '#F9F7F9'}}>
-          <article className="tile is-child" style={{backgroundColor: '#F9F7F9'}}>
+      <div className='tile is-parent is-vertical' style={{backgroundColor: 'transparent'}}>
+          <article className="tile is-child" style={{backgroundColor: 'transparent'}}>
             <p className="title is-2">Open Competitions</p>
             <div style={{display: 'flex', flexDirection: 'row'}}>
               <CompetitionCard />
@@ -41,7 +41,7 @@ class CompetitionRecord extends Component {
               <CompetitionCard />
             </div>
           </article>
-          <article className="tile is-child" style={{backgroundColor: '#F9F7F9'}}>
+          <article className="tile is-child" style={{backgroundColor: 'transparent'}}>
             <p className="title is-2">Past Competitions</p>
             <div style={{display: 'flex', flexDirection: 'row'}}>
               <CompetitionCard />
