+++ frontend/views/CompanyProfile.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
+import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 
 import CompanyProfileEdit from './CompanyProfile/CompanyProfileEdit';
 import CompanyProfilePublic from './CompanyProfile/CompanyProfilePublic';
@@ -52,9 +53,16 @@ class CompanyProfile extends Component {
 
   render() {
     return (
+      <ReactCSSTransitionGroup
+        transitionName="smallfade"
+        transitionAppear={true}
+        transitionAppearTimeout={350}
+        transitionEnter={false}
+        transitionLeave={false}>
       <div>
         {this.isEditing()}
       </div>
+    </ReactCSSTransitionGroup>
     )
   }
 }
