+++ frontend/views/CompanyProfile/CompanyProfileEdit.js
@@ -16,14 +16,8 @@ class CompanyProfileEdit extends Component {
   componentWillMount() {
     var html = document.getElementsByTagName("html")[0];
     var body = document.getElementsByTagName("body")[0];
-    $(html).attr("style", 'background-color: #6190E8');
-    $(body).attr("style", 'background-color: #6190E8')
-  }
-  componentWillUnmount() {
-    var html = document.getElementsByTagName("html")[0];
-    var body = document.getElementsByTagName("body")[0];
-    $(html).attr("style", 'background-color: #F9F7F9');
-    $(body).attr("style", 'background-color: #F9F7F9')
+    $(html).attr("style", 'background-color: #6190E8; transition:all ease-in-out 0.2s;');
+    $(body).attr("style", 'background-color: #6190E8; transition:all ease-in-out 0.2s;');
   }
 
   render() {
