+++ frontend/views/CompanyProfile/CompanyProfilePublic.js
@@ -15,14 +15,16 @@ class CompanyProfilePublic extends Component {
   componentWillMount() {
     var html = document.getElementsByTagName("html")[0];
     var body = document.getElementsByTagName("body")[0];
-    $(html).attr("style", 'background-color: #F9F7F9');
-    $(body).attr("style", 'background-color: #F9F7F9')
+    $(html).attr("style", 'background-color: #F9F7F9; transition:all ease-in-out 0.2s');
+    $(body).attr("style", 'background-color: #F9F7F9; transition:all ease-in-out 0.2s');
   }
 
   render() {
     const { username, email, pastCompetitions, currentCompetitions, website, type } = this.props.company;
 
     return(
+      <div>
+        {/* <div style={{position: 'absolute', top: '0', backgroundColor: 'red', height: '100%', width: '100%', zIndex: '0'}}></div> */}
       <div className="container is-fluid">
         <div className="top-spacer">
           {(this.props.id === this.props.user._id) &&
@@ -54,6 +56,7 @@ class CompanyProfilePublic extends Component {
           </div>
         </div>
       </div>
+    </div>
     )
 
   }
