+++ frontend/views/StudentProfile/StudentProfileEditable.js
@@ -21,14 +21,14 @@ class StudentProfileEditable extends Component {
   componentWillMount() {
     var html = document.getElementsByTagName("html")[0];
     var body = document.getElementsByTagName("body")[0];
-    $(html).attr("style", 'background-color: #6190E8');
-    $(body).attr("style", 'background-color: #6190E8')
+    $(html).attr("style", 'background-color: #6190E8; transition:all ease-in-out 0.2s;');
+    $(body).attr("style", 'background-color: #6190E8; transition:all ease-in-out 0.2s;');
   }
   componentWillUnmount() {
     var html = document.getElementsByTagName("html")[0];
     var body = document.getElementsByTagName("body")[0];
-    $(html).attr("style", 'background-color: #F9F7F9');
-    $(body).attr("style", 'background-color: #F9F7F9')
+    $(html).attr("style", 'background-color: #F9F7F9; transition:all ease-in-out 0.2s');
+    $(body).attr("style", 'background-color: #F9F7F9; transition:all ease-in-out 0.2s');
   }
   render() {
     const { username, email, fname, lname, university, year, major, competitionExperience,
