+++ public/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html style="background-color: #F9F7F9">
+<html class="htmlbg">
   <head>
     <meta charset="utf-8">
     <title>Mirathon</title>
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
     <!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.1/css/materialize.min.css"> -->
   </head>
-  <body>
+  <body class="htmlbg">
     <div id='root'/>
   </body>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
